requirements.txt
fastapi
uvicorn[standard]
pydantic>=2,<3
python-dotenv
jinja2
mangum
pytest
httpx
python-pptx
python-docx
playwright
pytest-playwright
PyYAML>=6.0
xlsxwriter
pdfplumber>=0.10.0
openpyxl>=3.1.0
PyJWT>=2.8.0
bcrypt>=4.0.0
cryptography>=42.0.0
ldap3>=2.9.0
python-multipart>=0.0.9
boto3>=1.34.0
psutil>=5.9.0
Markdown>=3.5.0
prometheus-client>=0.20.0
